arquitetura/pub-sub/notification-app/requirements.txt
kafka-python==2.0.2
requests==2.28.1
smtplib==0.1